// src/services/generated-client/models/SessionInitRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dual-System Digital Twin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Request payload for initializing a simulation session.
 * 
 * Attributes:
 *     model_name (str): The name of the model to simulate (e.g.,
 *         'water_tank').
 *     params (dict, optional): Model-specific initialization parameters.
 * @export
 * @interface SessionInitRequest
 */
export interface SessionInitRequest {
    /**
     * 
     * @type {string}
     * @memberof SessionInitRequest
     */
    modelName: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof SessionInitRequest
     */
    params?: { [key: string]: any; } | null;
}

/**
 * Check if a given object implements the SessionInitRequest interface.
 */
export function instanceOfSessionInitRequest(value: object): value is SessionInitRequest {
    if (!('modelName' in value) || value['modelName'] === undefined) return false;
    return true;
}

export function SessionInitRequestFromJSON(json: any): SessionInitRequest {
    return SessionInitRequestFromJSONTyped(json, false);
}

export function SessionInitRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionInitRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'modelName': json['model_name'],
        'params': json['params'] == null ? undefined : json['params'],
    };
}

export function SessionInitRequestToJSON(json: any): SessionInitRequest {
    return SessionInitRequestToJSONTyped(json, false);
}

export function SessionInitRequestToJSONTyped(value?: SessionInitRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'model_name': value['modelName'],
        'params': value['params'],
    };
}