// src/services/generated-client/models/SessionInitResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dual-System Digital Twin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Response payload containing the session ID for a new simulation.
 * @export
 * @interface SessionInitResponse
 */
export interface SessionInitResponse {
    /**
     * 
     * @type {string}
     * @memberof SessionInitResponse
     */
    sessionId: string;
}

/**
 * Check if a given object implements the SessionInitResponse interface.
 */
export function instanceOfSessionInitResponse(value: object): value is SessionInitResponse {
    if (!('sessionId' in value) || value['sessionId'] === undefined) return false;
    return true;
}

export function SessionInitResponseFromJSON(json: any): SessionInitResponse {
    return SessionInitResponseFromJSONTyped(json, false);
}

export function SessionInitResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionInitResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'sessionId': json['session_id'],
    };
}

export function SessionInitResponseToJSON(json: any): SessionInitResponse {
    return SessionInitResponseToJSONTyped(json, false);
}

export function SessionInitResponseToJSONTyped(value?: SessionInitResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'session_id': value['sessionId'],
    };
}